requirements.txt
Django==4.2
Flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.0
azure-ai-language-questionanswering
azure-ai-language-conversations
azure-identity==1.14.0
flask-cors==4.0.0
